chore: separate basis and theme validations into separate files
