+++ packages/design-tokens-schema/src/basis-tokens.test.ts
@@ -1,22 +1,5 @@
-import maTokens from '@nl-design-system-community/ma-design-tokens/dist/tokens';
-import startTokens from '@nl-design-system-unstable/start-design-tokens/dist/tokens';
-import voorbeeldTokens from '@nl-design-system-unstable/voorbeeld-design-tokens/dist/tokens';
 import { describe, it, expect } from 'vitest';
-import * as z from 'zod';
-import {
-  BrandsSchema,
-  BrandSchema,
-  BasisTokensSchema,
-  BasisColorSchema,
-  EXTENSION_CONTRAST_WITH,
-  EXTENSION_RESOLVED_AS,
-  Theme,
-  StrictThemeSchema,
-  ThemeSchema,
-  type ContrastExtension,
-} from './basis-tokens';
-import { ColorToken, parseColor } from './tokens/color-token';
-import { ERROR_CODES, type ThemeValidationIssue } from './validation-issue';
+import { BrandsSchema, BrandSchema, BasisTokensSchema, BasisColorSchema } from './basis-tokens';
 
 describe('brand', () => {
   it('no brands present', () => {
@@ -252,648 +235,5 @@ describe('basis', () => {
       const result = BasisColorSchema.safeParse(config);
       expect.soft(result.success).toEqual(true);
     });
-  });
-});
-
-describe('theme', () => {
-  const brandConfig = {
-    ma: {
-      name: {
-        $type: 'text',
-        $value: 'Mooi & Anders',
-      },
-      color: {
-        indigo: {
-          '1': {
-            $type: 'color',
-            $value: {
-              alpha: 1,
-              colorSpace: 'srgb',
-              components: [1, 1, 1],
-            },
-          },
-          '5': {
-            $type: 'color',
-            $value: {
-              alpha: 1,
-              colorSpace: 'srgb',
-              components: [0, 0, 0],
-            },
-          },
-        },
-      },
-    },
-  };
-
-  describe('adding contrast-with extensions', () => {
-    it('adds contrast-with extensions for tokens that we know', () => {
-      const config = {
-        basis: {
-          color: {
-            default: {
-              'bg-subtle': {
-                $type: 'color',
-                $value: '{ma.color.indigo.1}',
-              },
-              'color-document': {
-                $type: 'color',
-                $value: `{ma.color.indigo.5}`,
-                // the transformation will add an $extension here
-              },
-            },
-          },
-        },
-        brand: brandConfig,
-      };
-      const result = StrictThemeSchema.safeParse(config);
-      expect(result.success).toEqual(true);
-      expect(result.data?.basis?.color?.default?.['color-document']?.$extensions).toMatchObject({
-        [EXTENSION_CONTRAST_WITH]: [
-          {
-            color: {
-              $extensions: {
-                [EXTENSION_RESOLVED_AS]: config.brand.ma.color.indigo['1'].$value,
-              },
-              $type: 'color',
-              $value: '{ma.color.indigo.1}',
-            },
-            expectedRatio: 4.5,
-          },
-        ],
-      });
-    });
-
-    it('contrast-with extensions do not mess with existing extensions', () => {
-      const config = {
-        basis: {
-          color: {
-            default: {
-              'bg-subtle': {
-                $type: 'color',
-                $value: '{ma.color.indigo.1}',
-              },
-              'color-document': {
-                $extensions: {
-                  // the transformation will add an extra contrast color here
-                  [EXTENSION_CONTRAST_WITH]: [
-                    {
-                      color: {
-                        $type: 'color',
-                        $value: '{ma.color.indigo.5}',
-                      },
-                      ratio: 1,
-                    },
-                  ],
-                },
-                $type: 'color',
-                $value: `{ma.color.indigo.5}`,
-              },
-            },
-          },
-        },
-        brand: brandConfig,
-      };
-      const result = StrictThemeSchema.safeParse(config);
-      expect(result.success).toEqual(true);
-      expect(
-        result.data?.basis?.color?.default?.['color-document']?.$extensions?.[EXTENSION_CONTRAST_WITH],
-      ).toHaveLength(2);
-    });
-
-    it('does not add extension when corresponding token does not exist', () => {
-      const config = {
-        basis: {
-          color: {
-            default: {
-              'color-document': {
-                $type: 'color',
-                $value: `{ma.color.indigo.5}`,
-                // contrast-with extension would be added here but this object has no bg-subtle
-              },
-            },
-          },
-        },
-        brand: brandConfig,
-      };
-      const result = StrictThemeSchema.safeParse(config);
-      expect(result.success).toEqual(true);
-      expect(result.data?.basis?.color?.default?.['color-document']?.$extensions?.[EXTENSION_CONTRAST_WITH]).toEqual(
-        undefined,
-      );
-    });
-  });
-
-  describe('resolving Design Token refs', () => {
-    describe('resolve color ref', () => {
-      const config = {
-        basis: {
-          color: {
-            default: {
-              'bg-document': {
-                $type: 'color',
-                $value: `{ma.color.indigo.5}`,
-              },
-            },
-          },
-        },
-        brand: brandConfig,
-      };
-
-      it('does not mutate the input config', () => {
-        const originalConfig = structuredClone(config);
-        StrictThemeSchema.safeParse(config);
-        const originalBg = originalConfig.basis.color.default['bg-document'].$value;
-        const bgAfterValidation = config.basis.color.default['bg-document'].$value;
-        expect.soft(originalBg).toEqual(bgAfterValidation);
-      });
-
-      it('validates the input', () => {
-        const result = StrictThemeSchema.safeParse(config);
-        expect.soft(BrandSchema.safeParse(config.brand.ma).success).toBeTruthy();
-        expect.soft(result.success).toBeTruthy();
-      });
-
-      it('returns the schema with the ref value added to an extension and the $value left intact', () => {
-        const result = StrictThemeSchema.safeParse(config);
-        const expectedColor = brandConfig.ma.color.indigo[5];
-        expect.soft(result.data?.basis?.color?.default?.['bg-document']).toEqual({
-          ...config.basis.color.default['bg-document'],
-          $extensions: {
-            [EXTENSION_RESOLVED_AS]: expectedColor.$value,
-          },
-        });
-      });
-    });
-
-    it('marks as invalid if resolving to an nonexistent object', () => {
-      const config = {
-        basis: {
-          color: {
-            default: {
-              'bg-document': {
-                $type: 'color',
-                $value: `{non.existent.token}`,
-              },
-              'bg-subtle': {
-                $type: 'color',
-                $value: `{incomplete.ref`,
-              },
-            },
-          },
-        },
-        brand: brandConfig,
-      };
-
-      expect.soft(() => StrictThemeSchema.safeParse(config)).not.toThrowError();
-      const result = StrictThemeSchema.safeParse(config);
-      expect.soft(result.success).toBeFalsy();
-    });
-
-    it('marks as invalid if resolving to an existing object without a $value property', () => {
-      const config = {
-        basis: {
-          color: {
-            default: {
-              'bg-document': {
-                $type: 'color',
-                $value: `{ma.color.indigo}`,
-              },
-            },
-          },
-        },
-        brand: brandConfig,
-      };
-
-      expect(() => StrictThemeSchema.safeParse(config)).not.toThrowError();
-      const result = StrictThemeSchema.safeParse(config);
-      expect(result.success).toBeFalsy();
-      expect.soft(z.flattenError(result.error!)).toMatchObject({
-        formErrors: [
-          'Invalid token reference: expected "{ma.color.indigo}" to have a "$value" and "$type" property (referenced from "basis.color.default.bg-document")',
-        ],
-      });
-    });
-
-    it('marks as invalid when a font-family token reference points to an existing non-font-family token', () => {
-      const config = {
-        basis: {
-          heading: {
-            'font-family': {
-              $type: 'fontFamily',
-              // this points to a color token instead of a fontFamily
-              $value: '{ma.color.indigo.5}',
-            },
-          },
-        },
-        brand: brandConfig,
-      };
-
-      const result = StrictThemeSchema.safeParse(config);
-      expect.soft(result.success).toEqual(false);
-      expect.soft(z.flattenError(result.error!)).toMatchObject({
-        formErrors: [
-          `Invalid token reference: $type "fontFamily" of "{"$type":"fontFamily","$value":"{ma.color.indigo.5}"}" does not match the $type on reference {ma.color.indigo.5}. Types "fontFamily" and "color" do not match.`,
-        ],
-      });
-    });
-  });
-
-  describe('Style Dictionary specifics', () => {
-    const config = {
-      basis: {
-        color: {
-          'accent-1': {
-            'bg-default': {
-              $type: 'color',
-              $value: '#ffffff',
-              original: {
-                $type: 'color',
-                $value: '{ma.color.white}',
-              },
-            },
-          },
-        },
-      },
-      ma: {
-        color: {
-          white: {
-            $type: 'color',
-            $value: '#ffffff',
-          },
-        },
-      },
-    };
-
-    it('replaces token.$value with token.original.$value', () => {
-      const result = ThemeSchema.safeParse(config);
-      expect(result.success).toBe(true);
-      expect((result.data as Theme)?.basis?.color?.['accent-1']?.['bg-default']?.$value).toBe('{ma.color.white}');
-    });
-
-    it('ignores incomplete `original` objects', () => {
-      const incompleteConfig = structuredClone(config);
-      // @ts-expect-error what we're doing here is unexpeced, type-wise
-      delete incompleteConfig.basis.color['accent-1']['bg-default'].original.$value;
-      const result = ThemeSchema.safeParse(incompleteConfig);
-      expect(result.success).toBe(true);
-      expect((result.data as Theme)?.basis?.color?.['accent-1']?.['bg-default']?.$value).toEqual({
-        alpha: 1,
-        colorSpace: 'srgb',
-        components: [1, 1, 1],
-      });
-    });
-
-    it('adds resolved-as extension in StrictTheme', () => {
-      const result = StrictThemeSchema.safeParse(config);
-      expect(result.success).toBe(true);
-      expect(result.data?.basis?.color?.['accent-1']?.['bg-default']?.$value).toBe('{ma.color.white}');
-      expect(result.data?.basis?.color?.['accent-1']?.['bg-default']?.$extensions?.[EXTENSION_RESOLVED_AS]).toEqual({
-        alpha: 1,
-        colorSpace: 'srgb',
-        components: [1, 1, 1],
-      });
-    });
-  });
-
-  describe('validating color contrast', () => {
-    // parseColor() because it's shorter than writing color tokens and the IDE shows a small color preview
-    const black = { $type: 'color', $value: parseColor('#000') } satisfies ColorToken;
-    const white = { $type: 'color', $value: parseColor('#fff') } satisfies ColorToken;
-    const lightGray = { $type: 'color', $value: parseColor('#ccc') } satisfies ColorToken;
-    const config = {
-      basis: {
-        color: {
-          // This is where each test case will write their own needs
-          default: Object.create(null),
-        },
-      },
-      brand: {
-        ma: {
-          name: {
-            $type: 'text',
-            $value: 'Mooi & Anders',
-          },
-          color: {
-            black,
-            gray: {
-              2: lightGray,
-            },
-            white,
-          },
-        },
-      },
-    } satisfies Theme;
-
-    it('passes when contrast is sufficient', () => {
-      const testConfig = structuredClone(config);
-      testConfig.basis.color.default['border-default'] = {
-        $type: 'color',
-        $value: '{ma.color.black}',
-      };
-      testConfig.basis.color.default['bg-default'] = {
-        $type: 'color',
-        $value: '{ma.color.white}',
-      };
-      const result = StrictThemeSchema.safeParse(testConfig);
-      expect(result.success).toEqual(true);
-    });
-
-    it('fails when color-document vs bg-subtle do not have enough contrast', () => {
-      const testConfig = structuredClone(config);
-      testConfig.basis.color.default['bg-subtle'] = {
-        $type: 'color',
-        $value: '{ma.color.white}',
-      };
-      testConfig.basis.color.default['color-document'] = {
-        $type: 'color',
-        $value: '{ma.color.gray.2}',
-      };
-      const result = StrictThemeSchema.safeParse(testConfig);
-      expect(result.success).toBeFalsy();
-      expect.soft(result.error!.issues).toEqual([
-        {
-          actual: 1.6059285649300712,
-          code: 'too_small',
-          ERROR_CODE: 'insufficient_contrast',
-          message: 'Insufficient contrast',
-          minimum: 4.5,
-          origin: 'number',
-          path: 'basis.color.default.color-document.$value'.split('.'),
-          tokens: ['basis.color.default.color-document', 'basis.color.default.bg-subtle'],
-        },
-        {
-          actual: 1.6059285649300712,
-          code: 'too_small',
-          ERROR_CODE: 'insufficient_contrast',
-          message: 'Insufficient contrast',
-          minimum: 4.5,
-          origin: 'number',
-          path: 'basis.color.default.bg-subtle.$value'.split('.'),
-          tokens: ['basis.color.default.bg-subtle', 'basis.color.default.color-document'],
-        },
-      ]);
-    });
-
-    it('passes when legacy tokens are used with proper contrast', () => {
-      const testConfig = structuredClone(config);
-      testConfig.basis.color.default['bg-subtle'] = {
-        $type: 'color',
-        $value: '#fff',
-      };
-      testConfig.basis.color.default['color-document'] = {
-        $type: 'color',
-        $value: '#000',
-      };
-      const result = StrictThemeSchema.safeParse(testConfig);
-      expect(result.success).toEqual(true);
-    });
-
-    it('fails when legacy tokens are used with improper contrast', () => {
-      const testConfig = structuredClone(config);
-      testConfig.basis.color.default['bg-subtle'] = {
-        $type: 'color',
-        $value: '#fff',
-      };
-      testConfig.basis.color.default['color-document'] = {
-        $type: 'color',
-        $value: '#ccc',
-      };
-      const result = StrictThemeSchema.safeParse(testConfig);
-      expect(result.success).toEqual(false);
-      expect(result.error!.issues).toEqual([
-        {
-          actual: 1.6059285649300712,
-          code: 'too_small',
-          ERROR_CODE: 'insufficient_contrast',
-          message: 'Insufficient contrast',
-          minimum: 4.5,
-          origin: 'number',
-          path: 'basis.color.default.color-document.$value'.split('.'),
-          tokens: ['basis.color.default.color-document', 'basis.color.default.bg-subtle'],
-        },
-        {
-          actual: 1.6059285649300712,
-          code: 'too_small',
-          ERROR_CODE: 'insufficient_contrast',
-          message: 'Insufficient contrast',
-          minimum: 4.5,
-          origin: 'number',
-          path: 'basis.color.default.bg-subtle.$value'.split('.'),
-          tokens: ['basis.color.default.bg-subtle', 'basis.color.default.color-document'],
-        },
-      ]);
-    });
-
-    it('handles contrast error when background color has no EXTENSION_RESOLVED_FROM', () => {
-      const testConfig = structuredClone(config);
-      testConfig.basis.color.default['color-document'] = {
-        $extensions: {
-          [EXTENSION_CONTRAST_WITH]: [
-            {
-              // Background color without EXTENSION_RESOLVED_FROM
-              color: {
-                $type: 'color',
-                $value: parseColor('#fff'),
-                // No EXTENSION_RESOLVED_FROM extension
-              },
-              expectedRatio: 4.5,
-            },
-          ],
-        },
-        $type: 'color',
-        $value: '#ccc',
-      };
-      const result = StrictThemeSchema.safeParse(testConfig);
-      expect(result.success).toBeFalsy();
-      const contrastErrors = (result.error!.issues as ThemeValidationIssue[]).filter(
-        (issue) => issue.ERROR_CODE === ERROR_CODES.INSUFFICIENT_CONTRAST,
-      );
-      expect(contrastErrors.length).toBeGreaterThan(0);
-
-      const errorWithUndefinedPath = contrastErrors.find(
-        (error) => error.path?.length === 1 && error.path[0] === '$value',
-      );
-      expect(errorWithUndefinedPath).toBeDefined();
-    });
-  });
-
-  it('finds both ref and contrast issues at once', () => {
-    // This test case exists because previous versions would bail out after the validation found an invalid ref and skipped checking contrast
-    const white = parseColor('#ffffff');
-    const lightGray = parseColor('#cccccc');
-
-    const themeWithBothErrors: Theme = {
-      basis: {
-        color: {
-          default: {
-            // Invalid ref - this token doesn't exist
-            'bg-document': {
-              $type: 'color',
-              $value: '{test.color.nonexistent}',
-            },
-            // Insufficient contrast - gray on white is too low
-            'bg-subtle': {
-              $type: 'color',
-              $value: '{test.color.white}',
-            },
-            'color-document': {
-              $type: 'color',
-              $value: '{test.color.gray}',
-            },
-          },
-        },
-      },
-      brand: {
-        test: {
-          name: { $type: 'text', $value: 'Test' },
-          color: {
-            gray: { $type: 'color', $value: lightGray },
-            white: { $type: 'color', $value: white },
-          },
-        },
-      },
-    };
-
-    const result = StrictThemeSchema.safeParse(themeWithBothErrors);
-    expect(result.success).toBe(false);
-
-    // Should find both types of errors
-    const refErrors = (result.error!.issues as ThemeValidationIssue[]).filter(
-      (issue) => issue.ERROR_CODE === ERROR_CODES.INVALID_REF,
-    );
-    const contrastErrors = (result.error!.issues as ThemeValidationIssue[]).filter(
-      (issue) => issue.ERROR_CODE === ERROR_CODES.INSUFFICIENT_CONTRAST,
-    );
-
-    expect(refErrors.length).toBeGreaterThan(0);
-    expect(contrastErrors.length).toBeGreaterThan(0);
-  });
-
-  describe('remove non-token properties', () => {
-    it('removes Style Dictionary metadata', () => {
-      const theme = {
-        basis: {
-          color: {
-            default: {
-              'bg-document': {
-                $type: 'color',
-                $value: {
-                  colorSpace: 'srgb',
-                  components: [0, 0, 0],
-                },
-                // object
-                attributes: {
-                  category: 'basis',
-                  item: 'default',
-                  type: 'color',
-                },
-                // string
-                filePath: 'tokens.json',
-                // boolean
-                isSource: true,
-                // object with tokens inside
-                original: {
-                  $type: 'color',
-                  $value: {
-                    colorSpace: 'srgb',
-                    components: [0, 0, 0],
-                  },
-                },
-              },
-            },
-          },
-        },
-      } satisfies Theme;
-      const result = StrictThemeSchema.safeParse(theme);
-      expect(result.success).toBeTruthy();
-      const actual = result.data?.basis?.color?.default?.['bg-document'];
-      expect(actual?.$type).toBe('color');
-      expect(actual?.$value).toEqual({
-        colorSpace: 'srgb',
-        components: [0, 0, 0],
-      });
-      const actualAsRecord = actual as Record<string, unknown>;
-      expect(actualAsRecord?.['attributes']).toBeUndefined();
-      expect(actualAsRecord?.['filePath']).toBeUndefined();
-      expect(actualAsRecord?.['isSource']).toBeUndefined();
-      expect(actualAsRecord?.['original']).toBeUndefined();
-    });
-
-    it('does not remove information from $extensions (which could hold any sort of data)', () => {
-      const theme = {
-        basis: {
-          color: {
-            default: {
-              'bg-document': {
-                $extensions: {
-                  [EXTENSION_CONTRAST_WITH]: [
-                    {
-                      color: {
-                        $type: 'color',
-                        $value: parseColor('#fff'),
-                      },
-                      expectedRatio: 4.5,
-                    },
-                  ],
-                },
-                $type: 'color',
-                $value: {
-                  colorSpace: 'srgb',
-                  components: [0, 0, 0],
-                },
-              },
-            },
-          },
-        },
-      } satisfies Theme;
-      const result = StrictThemeSchema.safeParse(theme);
-      expect(result.success).toBeTruthy();
-      const extensions = result.data?.basis?.color?.default?.['bg-document']?.$extensions;
-      expect(Array.isArray(extensions?.[EXTENSION_CONTRAST_WITH])).toBeTruthy();
-      expect(extensions?.[EXTENSION_CONTRAST_WITH]).toHaveLength(1);
-      const extension = (extensions?.[EXTENSION_CONTRAST_WITH] as ContrastExtension[])[0];
-      expect(extension).toBeDefined();
-      expect(extension.color.$type).toBe('color');
-      expect(extension.expectedRatio).toBe(4.5);
-    });
-
-    it('Should not remove top-level non-tokens (like brands)', () => {
-      const theme = {
-        ams: {
-          heading: {
-            color: {
-              $type: 'color',
-              $value: {
-                colorSpace: 'srgb',
-                components: [0, 0, 0],
-              },
-            },
-          },
-        },
-      } satisfies Theme;
-      const result = StrictThemeSchema.safeParse(theme);
-      expect(result.success).toBeTruthy();
-      expect(result.data).toEqual(theme);
-    });
-  });
-});
-
-describe('strictly validate known basis themes', () => {
-  it('Mooi & Anders theme', () => {
-    const result = StrictThemeSchema.safeParse(maTokens);
-    expect(result.success).toEqual(true);
-    expect(result.data).toMatchSnapshot();
-  });
-
-  it('Voorbeeld theme', () => {
-    const result = StrictThemeSchema.safeParse(voorbeeldTokens);
-    expect(result.success).toEqual(true);
-    expect(result.data).toMatchSnapshot();
-  });
-
-  it('Start theme', () => {
-    const result = StrictThemeSchema.safeParse(startTokens);
-    expect(result.success).toEqual(true);
-    expect(result.data).toMatchSnapshot();
   });
 });
