+++ packages/design-tokens-schema/src/index.ts
@@ -3,5 +3,6 @@ export * from './tokens/color-token';
 export * from './tokens/dimension-token';
 export * from './tokens/fontfamily-token';
 export * from './tokens/token-reference';
-export * from './basis-tokens';
 export * from './validation-issue';
+export * from './basis-tokens';
+export * from './theme';
